GenreController Status Codes And UnitTest Done

+++ H3CinemaBooking/H3CinemaBooking.API/Controllers/GenreController.cs
@@ -2,7 +2,7 @@
 using H3CinemaBooking.Repository.Models;
 using H3CinemaBooking.Repository.Repositories;
 using Microsoft.AspNetCore.Mvc;
-using System.Security.Cryptography;
+using System.Collections.Generic;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -19,49 +19,74 @@ namespace H3CinemaBooking.API.Controllers
             _movieGenreRepository = movieGenreRepository;
         }
 
-        //TODO: Make a Get all here
-        // GET: api/<MovieGenreController>
+        // GET: api/<GenreController>
         [HttpGet]
         public ActionResult<List<Genre>> GetAll()
         {
             var result = _movieGenreRepository.GetAll();
+            if (result == null || result.Count == 0)
+            {
+                return NoContent();
+            }
             return Ok(result);
         }
 
-        // GET api/<MovieGenreController>/id
+        // GET api/<GenreController>/id
         [HttpGet("{id}")]
         public ActionResult<Genre> GetByID(int id)
         {
-            var movie = _movieGenreRepository.GetById(id);
-            if (movie == null)
+            var genre = _movieGenreRepository.GetById(id);
+            if (genre == null)
             {
                 return NotFound();
             }
-            return Ok(movie);
+            return Ok(genre);
         }
 
-        // POST api/<MovieGenreController>
+        // POST api/<GenreController>
         [HttpPost]
-        public ActionResult<Genre> Post(Genre genre)
+        public ActionResult<Genre> Post([FromBody] Genre genre)
         {
+            if (genre == null)
+            {
+                return BadRequest("Genre data is required");
+            }
             _movieGenreRepository.Create(genre);
             return Ok(genre);
         }
 
-        //Update Api Movie with Genre
+        // PUT api/<GenreController>/5
         [HttpPut("{id}")]
-        public ActionResult Update(int id, Genre genre)
+        public ActionResult Update(int id, [FromBody] Genre genre)
         {
+            if (genre == null)
+            {
+                return BadRequest();
+            }
+
+            var existingGenre = _movieGenreRepository.GetById(id);
+            if (existingGenre == null)
+            {
+                return NotFound();
+            }
+
+            existingGenre.GenreName = genre.GenreName;
+
             _movieGenreRepository.UpdateByID(id, genre);
-            return Ok();
+            return Ok(existingGenre);
         }
 
-        // DELETE api/<MovieGenreController>/ID
+        // DELETE api/<GenreController>/5
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
         {
+            var genre = _movieGenreRepository.GetById(id);
+            if (genre == null)
+            {
+                return NotFound();
+            }
             _movieGenreRepository.DeleteGenreByID(id);
-            return Ok();
+            return Ok(new { message = "Genre deleted successfully" });
         }
     }
 }
